pyproject.toml
[project]
name = "fla-dit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "diffusers>=0.35.2",
    "flash-linear-attention>=0.4.0",
    "matplotlib>=3.10.7",
    "transformers>=4.57.3",
    "webdataset>=1.0.2",
]
